Move navigator interface to features package
