+++ content-scope-scripts/content-scope-scripts-impl/src/main/java/com/duckduckgo/contentscopescripts/impl/features/navigatorinterface/store/NavigatorInterfaceDatabase.kt
@@ -14,11 +14,21 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.contentscopescripts.impl.navigatorinterface
+package com.duckduckgo.contentscopescripts.impl.features.navigatorinterface.store
 
-/**
- * Convenience method to get the [NavigatorInterfaceFeatureName] from its [String] value
- */
-fun navigatorInterfaceFeatureValueOf(value: String): NavigatorInterfaceFeatureName? {
-    return NavigatorInterfaceFeatureName.entries.find { it.value == value }
+import androidx.room.Database
+import androidx.room.RoomDatabase
+import androidx.room.migration.Migration
+
+@Database(
+    exportSchema = true,
+    version = 1,
+    entities = [
+        NavigatorInterfaceEntity::class,
+    ],
+)
+abstract class NavigatorInterfaceDatabase : RoomDatabase() {
+    abstract fun navigatorInterfaceDao(): NavigatorInterfaceDao
 }
+
+val ALL_MIGRATIONS = emptyArray<Migration>()
